Fixed bugs with tx handling

+++ trunk/hub/composites/push/src/main/java/se/vgregion/pubsub/push/impl/DefaultPushSubscriberManager.java
@@ -1,12 +1,16 @@
 package se.vgregion.pubsub.push.impl;
 
+import java.io.IOException;
+import java.net.URI;
 import java.util.Collection;
 
 import org.springframework.transaction.annotation.Transactional;
 
 import se.vgregion.pubsub.PubSubEngine;
 import se.vgregion.pubsub.Topic;
+import se.vgregion.pubsub.push.FailedSubscriberVerificationException;
 import se.vgregion.pubsub.push.PushSubscriber;
+import se.vgregion.pubsub.push.SubscriptionMode;
 import se.vgregion.pubsub.push.repository.PushSubscriberRepository;
 
 public class DefaultPushSubscriberManager implements PushSubscriberManager {
@@ -32,26 +36,33 @@ public class DefaultPushSubscriberManager implements PushSubscriberManager {
 
     @Override
     @Transactional
-    public void subscribe(PushSubscriber subscriber) {
-        unsubscribe(subscriber);
+    public void subscribe(URI topicUrl, URI callback, int leaseSeconds, String verifyToken) {
+        unsubscribe(topicUrl, callback);
         
-        Topic topic = pubSubEngine.getOrCreateTopic(subscriber.getTopic());
+        Topic topic = pubSubEngine.getOrCreateTopic(topicUrl);
+        
+        DefaultPushSubscriber subscriber = new DefaultPushSubscriber(subscriptionRepository, topicUrl, callback, leaseSeconds, verifyToken);
         
         topic.addSubscriber(subscriber);
         
-        // TODO handle resubscriptions
         subscriptionRepository.persist(subscriber);
     }
 
     @Override
     @Transactional
-    public void unsubscribe(PushSubscriber subscriber) {
-        Topic topic = pubSubEngine.getOrCreateTopic(subscriber.getTopic());
-
-        topic.removeSubscriber(subscriber);
-        
-        subscriptionRepository.remove(subscriber);
+    public void unsubscribe(URI topicUrl, URI callback) {
+        PushSubscriber subscriber = subscriptionRepository.findByTopicAndCallback(topicUrl, callback);
+        if(subscriber != null) {
+            try {
+                subscriber.verify(SubscriptionMode.UNSUBSCRIBE);
+                Topic topic = pubSubEngine.getOrCreateTopic(subscriber.getTopic());
+                topic.removeSubscriber(subscriber);
+                subscriptionRepository.remove(subscriber);
+            } catch (IOException e) {
+                // ignore
+            } catch (FailedSubscriberVerificationException e) {
+                // ignore
+            }
+        }
     }
-    
-    
 }
